test: align BaseFormView tests
