+++ ui/src/components/BaseFormView/tests/BaseFormViewGrups.test.tsx
@@ -1,9 +1,9 @@
 import { render, screen, within } from '@testing-library/react';
 import React from 'react';
 
-import { setUnifiedConfig } from '../../util/util';
-import BaseFormView from './BaseFormView';
-import { getGlobalConfigMockModificationToGroupsConfig } from './BaseFormConfigMock';
+import { setUnifiedConfig } from '../../../util/util';
+import BaseFormView from '../BaseFormView';
+import { getGlobalConfigMockModificationToGroupsConfig } from './configMocks';
 
 const handleFormSubmit = jest.fn();
